bussystem/requirements.txt
asgiref==3.7.2
Django==4.2.8
django-ckeditor==6.7.1
django-ckeditor-5==0.2.11
django-js-asset==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
